requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://mirrors.aliyun.com/pypi/simple
aiohttp==3.7.4.post0
aliyun-python-sdk-core-v3==2.13.32
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-kms==2.15.0
amqp==5.0.6; python_version >= '3.6'
arrow==1.1.1
asgiref==3.4.1; python_version >= '3.6'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
automat==20.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
charset-normalizer==2.0.6; python_version >= '3'
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==3.4.8; python_version >= '3.6'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
diff-match-patch==20200713; python_version >= '2.7'
django-cors-headers==3.8.0
django-import-export==2.6.0
django-ipware==4.0.0
django-jsonfield==1.4.1
django-otp==1.1.1
django-requestlogs==0.2.4
django-scim2==0.17.0
django==3.2.7
djangorestframework==3.12.4
drf-extensions==0.7.1
drf-spectacular==0.14.0
elementpath==2.3.2; python_version >= '3.6'
et-xmlfile==1.1.0; python_version >= '3.6'
greenlet==1.1.1; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
hyperlink==21.0.0
idna==3.2
importlib-resources==5.2.2; python_version >= '3.6'
incremental==21.3.0
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.0.1; python_version >= '3.6'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
jsonschema==3.2.0
jwcrypto==1.0
kombu==5.1.0; python_version >= '3.6'
ldap3==2.9.1
ldaptor==21.2.0
markuppy==1.14
markupsafe==2.0.1; python_version >= '3.6'
multidict==5.1.0; python_version >= '3.6'
mysqlclient==2.0.3
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
oss2==2.15.0
passlib==1.7.4
pillow==8.2.0
prompt-toolkit==3.0.20; python_full_version >= '3.6.2'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycryptodome==3.10.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyopenssl==20.0.1
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyrsistent==0.18.0; python_version >= '3.6'
pysaml2==7.0.1
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
pyyaml==5.4.1
redis==3.5.3
requests==2.26.0
scim2-filter-parser==0.3.5
service-identity==21.1.0
six==1.16.0
sly==0.3
sqlalchemy==1.4.25
sqlparse==0.4.2; python_version >= '3.5'
tablib[html,ods,xls,xlsx,yaml]==3.0.0; python_version >= '3.6'
toml==0.10.2
twisted[tls]==21.7.0
typing-extensions==3.10.0.2
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.5
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
xmlschema==1.7.1; python_version >= '3.6'
yarl==1.6.3; python_version >= '3.6'
zipp==3.5.0; python_version < '3.10'
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
